add --avg and --std arguments and defaults for kallisto single-end reads
Previous versions of arcasHLA used the read length and standard deviation
which is the incorrect behaviour. It should use the fragment length and
standard deviation which must be determined experimentally.

The kallisto manual states: "Typical Illumina libraries produce fragment
lengths ranging from 180–200 bp but it’s best to determine this from a
library quantification with an instrument such as an Agilent Bioanalyzer."
and gives an example with "-l 200 -s 20". These values are thus chosen
as sane defaults in the absense of experimental info.

This commit should substantially improve the performance of the slow
and memory hungry analyze_reads.